requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
anthropic
openai
aiofiles
aiosqlite
python-dotenv
